+++ src/library/policy/upgrades.py
@@ -1,6 +1,10 @@
 # -*- coding: utf-8 -*-
+from eea.facetednavigation.subtypes.interfaces import IPossibleFacetedNavigable
 from plone import api
 from plone.app.upgrade.utils import loadMigrationProfile
+from zope.globalrequest import getRequest
+
+import os
 
 
 def reload_gs_profile(context):
@@ -22,3 +26,22 @@ def change_language(context):
         if obj.language != default_language:
             obj.language = default_language
     root.language = default_language
+
+
+def configure_faceted(context):
+    obj = api.portal.get()["explorer"]
+    if not IPossibleFacetedNavigable.providedBy(obj):
+        return
+    subtyper = obj.restrictedTraverse("@@faceted_subtyper")
+
+    if not subtyper:
+        return
+    subtyper.enable()
+    faceted_config_path = "{}/faceted/config/explorer.xml".format(
+        os.path.dirname(__file__)
+    )
+    with open(faceted_config_path, "rb") as faceted_config:
+        obj.unrestrictedTraverse("@@faceted_exportimport").import_xml(
+            import_file=faceted_config
+        )
+    request = getRequest()